- Modificati i tipi di alcune statistiche: ora sono tutte di tipo int - Aggiunti commenti ove necessario

+++ include/Monster.h
@@ -10,43 +10,58 @@
 
 class Monster {
     private:
-        double LP, DEF, ATK;
+        int LP, DEF, ATK;
         int LV;
         Point POS;
         std::string Name;
-        char symbol;   
+        char symbol;
 
     public:
 
     Monster();
+    //  Costruttore di default
 
     Monster(std::string Name, int LV);
+    //  Dati un nome e un intero che indica il livello, restituisce un mostro
 
-    double getLP();
+    int getLP();
+    //  Restituisce i punti vita del mostro
 
-    double getDEF();
+    int getDEF();
+    //  Restituisce i punti difesa del mostro
 
-    double getATK();
+    int getATK();
+    //  Restituisce i punti di attacco del mostro
 
     Point getPosition();
+    //  Restituisce la posizione del mostro
 
     int getLV();
+    //  Restituisce il livello del mostro
 
-    void setLP(double LP);
+    void setLP(int LP);
+    //  Modifica i punti vita del mostro
 
-    void setDEF(double DEF);
+    void setDEF(int DEF);
+    //  Modifica i punti difesa del mostro
 
-    void setATK(double ATK);
+    void setATK(int ATK);
+    //  Modifica i punti attacco del mostro
 
-    void setPosition(double x, double y);
+    void setPosition(int x, int y);
+    //  Date due coordinate x y, modifica la posizione del mostro
 
     std::string getName();
+    //  Restituisce il nome del mostro
 
     char getSymbol();
+    //  Restituisce il simbolo che indica il mostro
 
     friend bool operator==(Monster m1,Monster m2);
+    //  Definisce l'operatore "==" per la classe Monster
 
     friend bool operator!=(Monster m1,Monster m2);
+    //  Definisce l'operatore "!=" per la classe Monster
 };
 
 namespace std
@@ -54,7 +69,7 @@ namespace std
     template <> struct hash<Monster>
     {
         size_t operator()(Monster m) const;
-        // Hash function for the class Item
+        // Hash function for the class Monster
     };
 }
